+++ src/bot/commands/economy/public/bank.ts
@@ -1,0 +1,50 @@
+import { SlashCommandBuilder, ChatInputCommandInteraction, MessageFlags }
+  from "discord.js";
+
+import { sendSimpleEmbed, internalErrorEmbed }
+  from "../../../Helpers/simplified_embed_builder";
+
+import { getBalanceBW }
+  from "../../../services/database/tables/clients/manager";
+
+import { requireGuild }
+  from "../../../Helpers/require_guild";
+
+import { Command }
+  from "../../types";
+
+import { getEcoSymbol }
+  from "../../../services/database/tables/servers/configs/get_eco_symbol";
+
+export const getBankBalance: Command = {
+  data: new SlashCommandBuilder()
+    .setName('bank_balance')
+    .setDescription('see your bank balance')
+    .addBooleanOption(opt => opt
+      .setName('visibility')
+      .setDescription('Other people can see your balance')
+      .setRequired(true)
+    ),
+
+  async execute(interaction: ChatInputCommandInteraction) {
+    if (!(requireGuild(interaction))) return;
+    const isPublic = interaction.options.getString('visibility', true);
+    const guildId = interaction.guild!.id;
+    const userId = interaction.user.id;
+    const symbol = await getEcoSymbol(guildId)
+    await interaction.deferReply({ flags: !isPublic ? MessageFlags.Ephemeral : undefined })
+    try {
+      const userBal = await getBalanceBW(userId, guildId)
+      await sendSimpleEmbed(interaction, {
+        title: 'Bank balance 💳',
+        description: `${interaction.user} Your current balance is \`${symbol}${userBal}\``,
+        eph: !isPublic
+      }
+
+      )
+    } catch (e) {
+      console.log(e)
+      await internalErrorEmbed(interaction);
+    }
+  }
+}