+++ src/bot/commands/economy/public/wallet.ts
@@ -1,16 +1,16 @@
-import {SlashCommandBuilder, SlashCommandOptionsOnlyBuilder, MessageFlags, ChatInputCommandInteraction}
-from "discord.js";
+import { SlashCommandBuilder, SlashCommandOptionsOnlyBuilder, ChatInputCommandInteraction, MessageFlags }
+  from "discord.js";
 import { getBalanceBW }
-from "../../../services/database/tables/clients/manager";
+  from "../../../services/database/tables/clients/manager";
 
 import { internalErrorEmbed, sendSimpleEmbed }
-from "../../../Helpers/simplified_embed_builder";
+  from "../../../Helpers/simplified_embed_builder";
 
 import { getEcoSymbol }
-from "../../../services/database/tables/servers/configs/get_eco_symbol";
+  from "../../../services/database/tables/servers/configs/get_eco_symbol";
 
 import { requireGuild }
-from "../../../Helpers/require_guild";
+  from "../../../Helpers/require_guild";
 
 export interface Command {
   data: SlashCommandBuilder | SlashCommandOptionsOnlyBuilder;
@@ -30,22 +30,22 @@ export const getWalletBalance: Command = {
 
   async execute(interaction: ChatInputCommandInteraction) {
     if (!(await requireGuild(interaction))) return;
-    const visible = interaction.options.getBoolean("visibility", true);
+    const isPublic = interaction.options.getBoolean("visibility", true);
     const userId = interaction.user.id;
     const guildId = interaction.guild!.id;
-    await interaction.deferReply({flags: !visible ? MessageFlags.Ephemeral : undefined}); //^ Here we defer the message
-    try{
+    await interaction.deferReply({ flags: !isPublic ? MessageFlags.Ephemeral : undefined }); //^ Here we defer the message
+    try {
       const symbol = await getEcoSymbol(guildId);
       const userBal = await getBalanceBW(userId, guildId);
       await sendSimpleEmbed(interaction, {
         title: "Wallet Balance 💵",
-        description: `<@${userId}> Your current balance is \`${symbol}${userBal}\``,
-        eph: !visible,
+        description: `${interaction.user} Your current balance is \`${symbol}${userBal}\``,
+        eph: !isPublic,
       });
-    }catch(e){
+    } catch (e) {
       console.log(e)
       await internalErrorEmbed(interaction);
     }
-    
+
   },
 };